+++ src/main/java/com/example/datn/repository/ThuocTinhRepo/DuongKinhCanhRepo.java
@@ -8,9 +8,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
-
 @Repository
-public interface DuongKinhCanhRepo extends JpaRepository<DuongKinhCanh, Integer> {
+
+public interface DuongKinhCanhRepo extends JpaRepository<DuongKinhCanh, Long> {
     @Query("SELECT dkc FROM DuongKinhCanh dkc " +
             "WHERE (LOWER(dkc.ten) LIKE LOWER(CONCAT('%', :query, '%')) " +
             "AND (:trangThai IS NULL OR dkc.trangThai = :trangThai))")
