+++ src/main/java/com/example/datn/repository/ThuocTinhRepo/HangRepo.java
@@ -1,5 +1,6 @@
 package com.example.datn.repository.ThuocTinhRepo;
 
+
 import com.example.datn.entity.ThuocTinh.Hang;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -8,9 +9,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
-
 @Repository
-public interface HangRepo extends JpaRepository<Hang, Integer> {
+
+public interface HangRepo extends JpaRepository<Hang, Long> {
     @Query("SELECT h FROM Hang h " +
             "WHERE (LOWER(h.ten) LIKE LOWER(CONCAT('%', :query, '%')) " +
             "AND (:trangThai IS NULL OR h.trangThai = :trangThai))")
